pyproject.toml
[tool.poetry]
name = "income-expense-tracker"
version = "0.1.0"
description = "A web app for tracking incomes and expenses"
authors = ["Yuliia Shaposhnikova <domik339@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
pydantic = {extras = ["dotenv"], version = "^2.10.3"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
config = "^0.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
